Move kernel version utils to src/common
