+++ src/common/system/testdata/BUILD.bazel
@@ -23,6 +23,14 @@ filegroup(
     srcs = glob(include = ["proc/**/*"]),
 )
 
+filegroup(
+    name = "kernel_version",
+    srcs = glob([
+        "sample_host_ubuntu/**/*",
+        "sample_host_debian/**/*",
+    ]),
+)
+
 # For scoped_namespace_test: Generate a container with a dummy file inside,
 # so we can check for mount namespaces.
 genrule(
