+++ src/common/system/testdata/sample_host_debian/proc/sys/kernel/version
