+++ src/common/system/testdata/sample_host_ubuntu/proc/sys/kernel/version
