+++ src/common/system/testdata/sample_host_ubuntu/proc/version
