+++ src/stirling/utils/linux_headers.cc
@@ -18,15 +18,9 @@
 
 #include "src/stirling/utils/linux_headers.h"
 
-#include <elf.h>
-#include <link.h>
-#include <sys/auxv.h>
-#include <sys/utsname.h>
-
 #include <fstream>
 #include <limits>
 #include <memory>
-#include <sstream>
 #include <vector>
 
 #include <absl/strings/numbers.h>
@@ -49,235 +43,6 @@
 namespace px {
 namespace stirling {
 namespace utils {
-
-using px::system::ProcPath;
-using px::system::ProcPidRootPath;
-
-StatusOr<KernelVersion> ParseKernelVersionString(const std::string& linux_release_str) {
-  KernelVersion kernel_version;
-
-  static std::regex uname_regex("^([0-9]+)\\.([0-9]+)\\.([0-9]+)");
-  std::smatch match;
-  bool s = std::regex_search(linux_release_str, match, uname_regex);
-  if (!s || match.size() != 4) {
-    return error::Internal("Could not parse kernel version string: [status=$0, match_size=$1]", s,
-                           match.size());
-  }
-
-  uint32_t tmp;
-
-  // SimpleAtoi should always succeed, due to regex match.
-  ECHECK(absl::SimpleAtoi(match[1].str(), &tmp));
-  if (tmp > std::numeric_limits<uint16_t>::max()) {
-    return error::Internal("Kernel version does not appear valid %d", tmp);
-  }
-  kernel_version.version = tmp;
-
-  ECHECK(absl::SimpleAtoi(match[2].str(), &tmp));
-  if (tmp > std::numeric_limits<uint8_t>::max()) {
-    return error::Internal("Kernel major_rev does not appear valid %d", tmp);
-  }
-  kernel_version.major_rev = tmp;
-
-  ECHECK(absl::SimpleAtoi(match[3].str(), &tmp));
-  if (tmp > std::numeric_limits<uint8_t>::max()) {
-    return error::Internal("Kernel minor_rev does not appear valid %d", tmp);
-  }
-  kernel_version.minor_rev = tmp;
-
-  return kernel_version;
-}
-
-StatusOr<std::string> GetUname() {
-  // The following effectively runs `uname -r`.
-  struct utsname buffer;
-  if (uname(&buffer) != 0) {
-    LOG(ERROR) << "Could not determine kernel version";
-    return error::Internal("Could not determine kernel version (uname -r)");
-  }
-
-  std::string version_string(buffer.release);
-  LOG(INFO) << absl::Substitute("Obtained Linux version string from `uname`: $0", version_string);
-
-  return version_string;
-}
-
-StatusOr<std::string> GetProcVersionSignature() {
-  const auto version_signature_path = ProcPath("version_signature");
-  PX_ASSIGN_OR_RETURN(std::string version_signature, ReadFileToString(version_signature_path));
-
-  LOG(INFO) << absl::Substitute("Obtained Linux version string from $0: $1",
-                                version_signature_path.string(), version_signature);
-
-  // Example version signature:
-  // Ubuntu 4.15.0-96.97-generic 4.15.18
-
-  std::vector<std::string_view> tokens =
-      absl::StrSplit(version_signature, " ", absl::SkipWhitespace());
-  if (!tokens.empty()) {
-    return std::string(tokens.back());
-  }
-
-  return error::NotFound("Could not parse /proc/version_signature file");
-}
-
-StatusOr<std::string> GetProcSysKernelVersion() {
-  const auto proc_sys_kernel_version = ProcPath("sys", "kernel", "version");
-  PX_ASSIGN_OR_RETURN(const std::string version_string, ReadFileToString(proc_sys_kernel_version));
-
-  LOG(INFO) << absl::Substitute("Obtained Linux version string from $0: $1",
-                                proc_sys_kernel_version.string(), version_string);
-
-  // Example contents:
-  // #1 SMP Debian 4.19.152-1 (2020-10-18)
-
-  static std::regex version_regex(R"([0-9]+\.[0-9]+\.[0-9]+)");
-  std::smatch matches;
-  bool match_found = std::regex_search(version_string, matches, version_regex);
-  if (match_found) {
-    return std::string(matches[0]);
-  }
-
-  return error::NotFound("Could not find version number in /proc/sys/kernel/version file.");
-}
-
-// The Linux vDSO is a virtual dynamic shared object exposed by the Linux kernel.
-// The shared object is in ELF format and can contains metadata in a .note ELF section.
-// These notes--which are key, value pairs--can contain the Linux version code.
-// This function searches the vDSO to find a note about the Linux kernel version.
-StatusOr<KernelVersion> GetLinuxVersionFromNoteSection() {
-  // Get the vDSO object through getauxval().
-  // More information here: https://man7.org/linux/man-pages/man7/vdso.7.html
-  auto* vdso = reinterpret_cast<char*>(getauxval(AT_SYSINFO_EHDR));
-  if (vdso == nullptr) {
-    return error::NotFound("Could not find vDSO.");
-  }
-
-  if (!absl::StartsWith(vdso, ELFMAG)) {
-    return error::NotFound("vDSO does not appear to be ELF.");
-  }
-
-  // Get the ELF header.
-  auto elf_hdr = reinterpret_cast<const ElfW(Ehdr)*>(vdso);
-
-  // Go through all ELF section headers, looking for a .note section.
-  for (int i = 0; i < elf_hdr->e_shnum; ++i) {
-    auto sec_hdr =
-        reinterpret_cast<const ElfW(Shdr)*>(vdso + elf_hdr->e_shoff + (i * elf_hdr->e_shentsize));
-
-    // Look for a .note section.
-    if (sec_hdr->sh_type == SHT_NOTE) {
-      std::string_view note_data(reinterpret_cast<const char*>(vdso + sec_hdr->sh_offset),
-                                 sec_hdr->sh_size);
-
-      // Now prcess the .note data as {name, description} pairs.
-      // We search for a note that has name "Linux".
-      while (!note_data.empty()) {
-        auto notes_hdr = reinterpret_cast<const ElfW(Nhdr)*>(note_data.data());
-        note_data.remove_prefix(sizeof(const ElfW(Nhdr)));
-
-        std::string_view name(note_data.data(), notes_hdr->n_namesz);
-        note_data.remove_prefix(SnapUpToMultiple(name.size(), sizeof(ElfW(Word))));
-
-        std::string_view desc(note_data.data(), notes_hdr->n_descsz);
-        note_data.remove_prefix(SnapUpToMultiple(desc.size(), sizeof(ElfW(Word))));
-
-        // Strip off null terminator.
-        while (!name.empty() && name.back() == '\x00') {
-          name.remove_suffix(1);
-        }
-
-        if (name == "Linux" && desc.size() == 4 && notes_hdr->n_type == 0) {
-          uint32_t code = *reinterpret_cast<const uint32_t*>(desc.data());
-
-          uint8_t version = (code >> 16);
-          uint8_t major_rev = (code >> 8);
-          uint8_t minor_rev = (code >> 0);
-
-          if (code >> 24 != 0 || version < 1 || version > 9) {
-            return error::Internal("Linux version from vDSO appears corrupted.");
-          }
-
-          return KernelVersion{version, major_rev, minor_rev};
-        }
-      }
-    }
-  }
-
-  return error::NotFound("Could not extract kernel version from vDSO .note section.");
-}
-
-StatusOr<KernelVersion> FindKernelVersion(std::vector<KernelVersionSource> sources) {
-  for (const auto& source : sources) {
-    switch (source) {
-      // Use vDSO .note section to find Linux kernel version.
-      case KernelVersionSource::kVDSONoteSection: {
-        const StatusOr<KernelVersion> version = GetLinuxVersionFromNoteSection();
-        if (version.ok()) {
-          LOG(INFO) << "Found Linux kernel version using .note section.";
-          return version.ValueOrDie();
-        }
-      } break;
-
-      // Check /proc/version_signature.
-      // Required for Ubuntu distributions.
-      case KernelVersionSource::kProcVersionSignature: {
-        const StatusOr<std::string> version_string = GetProcVersionSignature();
-        if (version_string.ok()) {
-          LOG(INFO) << "Found Linux kernel version using /proc/version_signature.";
-          return ParseKernelVersionString(version_string.ValueOrDie());
-        }
-      } break;
-
-      // Use /proc/sys/kernel/version.
-      // Required for Debian distributions.
-      case KernelVersionSource::kProcSysKernelVersion: {
-        const StatusOr<std::string> version_string = GetProcSysKernelVersion();
-        if (version_string.ok()) {
-          LOG(INFO) << "Found Linux kernel version using /proc/sys/kernel/version.";
-          return ParseKernelVersionString(version_string.ValueOrDie());
-        }
-      } break;
-
-      // Use `uname -r`.
-      // Use this as a lower priority, because on Debian and Ubuntu,
-      // the uname does  not provide the correct minor version.
-      case KernelVersionSource::kUname: {
-        const StatusOr<std::string> version_string = GetUname();
-        if (version_string.ok()) {
-          LOG(INFO) << "Found Linux kernel version using uname.";
-          return ParseKernelVersionString(version_string.ValueOrDie());
-        }
-      } break;
-
-      default:
-        LOG(DFATAL) << absl::Substitute("Unhandled KernelVersionSource: $0",
-                                        magic_enum::enum_name(source));
-        break;
-    }
-  }
-
-  return error::Internal("Could not determine kernel version.");
-}
-
-StatusOr<KernelVersion> GetKernelVersion(std::vector<KernelVersionSource> sources) {
-  static std::optional<KernelVersion> g_kernel_version;
-
-  if (g_kernel_version.has_value()) {
-    return g_kernel_version.value();
-  }
-
-  auto version_or = FindKernelVersion(sources);
-  if (version_or.ok()) {
-    g_kernel_version = version_or.ValueOrDie();
-  }
-  return version_or;
-}
-
-KernelVersion GetCachedKernelVersion() {
-  static auto kernel_version = GetKernelVersion();
-  return kernel_version.ConsumeValueOr({0, 0, 0});
-}
 
 Status ModifyKernelVersion(const std::filesystem::path& linux_headers_base,
                            uint32_t linux_version_code) {
@@ -308,13 +73,13 @@ Status ModifyKernelVersion(const std::filesystem::path& linux_headers_base,
 //  - /lib/modules/<uname>/config: Used by RHEL8 CoreOS, and potentially other RHEL distros.
 StatusOr<std::filesystem::path> FindKernelConfig() {
   const system::Config& syscfg = system::Config::GetInstance();
-  PX_ASSIGN_OR_RETURN(const std::string uname, GetUname());
+  PX_ASSIGN_OR_RETURN(const std::string uname, system::GetUname());
 
   const std::vector<std::filesystem::path> search_paths = {
       // Used when CONFIG_IKCONFIG=y is set.
-      ProcPath("config"),
+      system::ProcPath("config"),
       // Used when CONFIG_IKCONFIG_PROC=y is set.
-      ProcPath("config.gz"),
+      system::ProcPath("config.gz"),
       // Search for /boot/config-<uname>
       syscfg.ToHostPath(absl::StrCat("/boot/config-", uname)),
       // Search for /lib/modules/<uname>/config
@@ -322,9 +87,9 @@ StatusOr<std::filesystem::path> FindKernelConfig() {
       // TODO(yzhao): https://github.com/lima-vm/alpine-lima/issues/67 once this issue is resolved,
       // we might consider change these 2 paths into something recommended by rancher-desktop.
       // The path used by `alpine-lima` in "Live CD" boot mechanism.
-      ProcPidRootPath(1, "media", "sr0", "boot", "config-virt"),
+      system::ProcPidRootPath(1, "media", "sr0", "boot", "config-virt"),
       // The path used by `alpine-lima` in "Live CD" boot mechanism on Mac machine.
-      ProcPidRootPath(1, "media", "vda", "boot", "config-virt"),
+      system::ProcPidRootPath(1, "media", "vda", "boot", "config-virt"),
   };
   std::vector<std::string> searched;
 
@@ -519,43 +284,8 @@ Status ExtractPackagedHeaders(const PackagedLinuxHeadersSpec& headers_package,
   return Status::OK();
 }
 
-uint64_t KernelHeadersDistance(KernelVersion a, KernelVersion b) {
-  // minor_rev range: 0-255
-  // major_rev range: 0-255
-  // version range: 0-65535
-
-  // Distance function makes sure major rev change is greater than 2x the minor rev change,
-  // so two versions on the same major rev are always closer than two version on different major
-  // revs. Similar logic for version.
-
-  return (abs(a.minor_rev - b.minor_rev)) + (abs(a.major_rev - b.major_rev) << 9) +
-         (abs(a.version - b.version) << 18);
-}
-
-KernelVersionOrder CompareKernelVersions(KernelVersion a, KernelVersion b) {
-  if (a.version < b.version) {
-    return KernelVersionOrder::kOlder;
-  }
-  if (a.version > b.version) {
-    return KernelVersionOrder::kNewer;
-  }
-  if (a.major_rev < b.major_rev) {
-    return KernelVersionOrder::kOlder;
-  }
-  if (a.major_rev > b.major_rev) {
-    return KernelVersionOrder::kNewer;
-  }
-  if (a.minor_rev < b.minor_rev) {
-    return KernelVersionOrder::kOlder;
-  }
-  if (a.minor_rev > b.minor_rev) {
-    return KernelVersionOrder::kNewer;
-  }
-  return KernelVersionOrder::kSame;
-}
-
 StatusOr<PackagedLinuxHeadersSpec> FindClosestPackagedLinuxHeaders(
-    const std::filesystem::path& packaged_headers_root, KernelVersion kernel_version) {
+    const std::filesystem::path& packaged_headers_root, system::KernelVersion kernel_version) {
 #if X86_64
   const std::string kHeaderDirPrefix =
       std::filesystem::path(packaged_headers_root / "linux-headers-x86_64-").string();
@@ -583,14 +313,14 @@ StatusOr<PackagedLinuxHeadersSpec> FindClosestPackagedLinuxHeaders(
       version_string_view.remove_suffix(kHeaderDirSuffix.size());
       std::string version_string(version_string_view);
 
-      StatusOr<KernelVersion> headers_kernel_version_status =
-          ParseKernelVersionString(version_string);
+      StatusOr<system::KernelVersion> headers_kernel_version_status =
+          system::ParseKernelVersionString(version_string);
       if (!headers_kernel_version_status.ok()) {
         LOG(WARNING) << absl::Substitute(
             "Ignoring $0 since it does not conform to the naming convention", path);
         continue;
       }
-      KernelVersion headers_kernel_version = headers_kernel_version_status.ValueOrDie();
+      system::KernelVersion headers_kernel_version = headers_kernel_version_status.ValueOrDie();
 
       if (selected.path.empty() || KernelHeadersDistance(kernel_version, headers_kernel_version) <
                                        KernelHeadersDistance(kernel_version, selected.version)) {
@@ -615,7 +345,7 @@ Status InstallPackagedLinuxHeaders(const std::filesystem::path& lib_modules_dir)
 
   LOG(INFO) << "Attempting to install packaged headers.";
 
-  PX_ASSIGN_OR_RETURN(KernelVersion kernel_version, GetKernelVersion());
+  PX_ASSIGN_OR_RETURN(system::KernelVersion kernel_version, system::GetKernelVersion());
 
   PX_ASSIGN_OR_RETURN(PackagedLinuxHeadersSpec packaged_headers,
                       FindClosestPackagedLinuxHeaders(kPackagedHeadersRoot, kernel_version));
@@ -651,7 +381,7 @@ Status InstallPackagedLinuxHeaders(const std::filesystem::path& lib_modules_dir)
 }
 
 Status FindOrInstallLinuxHeaders() {
-  PX_ASSIGN_OR_RETURN(std::string uname, GetUname());
+  PX_ASSIGN_OR_RETURN(std::string uname, system::GetUname());
   LOG(INFO) << absl::Substitute("Detected kernel release (uname -r): $0", uname);
 
   // BCC checks for Linux headers in both of the two following locations:
