+++ src/stirling/utils/linux_headers_test.cc
@@ -19,6 +19,7 @@
 #include <fstream>
 
 #include "src/common/system/config.h"
+#include "src/common/system/kernel_version.h"
 #include "src/common/testing/testing.h"
 #include "src/stirling/utils/linux_headers.h"
 
@@ -28,114 +29,11 @@ namespace px {
 namespace stirling {
 namespace utils {
 
-namespace {
-std::string GetPathToTestDataFile(const std::string_view fname) {
-  constexpr char kTestDataBasePath[] = "src/stirling/utils/testdata";
-  return testing::BazelRunfilePath(std::filesystem::path(kTestDataBasePath) / fname);
-}
-}  // namespace
-
 using ::px::testing::TempDir;
 using ::testing::EndsWith;
 
-bool operator==(const KernelVersion& a, const KernelVersion& b) {
+bool operator==(const system::KernelVersion& a, const system::KernelVersion& b) {
   return (a.version == b.version) && (a.major_rev == b.major_rev) && (a.minor_rev == b.minor_rev);
-}
-
-// A list of kernel version identification methods that use /proc only.
-// Used in a number of tests below.
-const std::vector<KernelVersionSource> kProcFSKernelVersionSources = {
-    KernelVersionSource::kProcVersionSignature, KernelVersionSource::kProcSysKernelVersion};
-
-TEST(LinuxHeadersUtils, LinuxVersionCode) {
-  KernelVersion version{4, 18, 1};
-  EXPECT_EQ(version.code(), 266753);
-}
-
-TEST(LinuxHeadersUtils, ParseKernelVersionString) {
-  EXPECT_OK_AND_EQ(ParseKernelVersionString("4.18.0-25-generic"), (KernelVersion{4, 18, 0}));
-  EXPECT_OK_AND_EQ(ParseKernelVersionString("4.18.4"), (KernelVersion{4, 18, 4}));
-  EXPECT_NOT_OK(ParseKernelVersionString("4.18."));
-  EXPECT_NOT_OK(ParseKernelVersionString("linux-4.18.0-25-generic"));
-}
-
-TEST(LinuxHeadersUtils, GetKernelVersionFromUname) {
-  StatusOr<KernelVersion> kernel_version_status = FindKernelVersion({KernelVersionSource::kUname});
-  ASSERT_OK(kernel_version_status);
-  KernelVersion kernel_version = kernel_version_status.ValueOrDie();
-
-  // We don't know on what host this test will run, so we don't know what the version code will be.
-  // But we can put some bounds, to check for obvious screw-ups.
-  // We assume test will run on a Linux machine with kernel 3.x.x or higher,
-  // and version 9.x.x or lower.
-  // Yes, we're being very generous here, but we want this test to pass the test of time.
-  EXPECT_GE(kernel_version.code(), 0x030000);
-  EXPECT_LE(kernel_version.code(), 0x090000);
-}
-
-TEST(LinuxHeadersUtils, GetKernelVersionFromNoteSection) {
-  StatusOr<KernelVersion> kernel_version_status =
-      FindKernelVersion({KernelVersionSource::kVDSONoteSection});
-  ASSERT_OK(kernel_version_status);
-  KernelVersion kernel_version = kernel_version_status.ValueOrDie();
-
-  // We don't know on what host this test will run, so we don't know what the version code will be.
-  // But we can put some bounds, to check for obvious screw-ups.
-  // We assume test will run on a Linux machine with kernel 3.x.x or higher,
-  // and version 9.x.x or lower.
-  // Yes, we're being very generous here, but we want this test to pass the test of time.
-  EXPECT_GE(kernel_version.code(), 0x030000);
-  EXPECT_LE(kernel_version.code(), 0x090000);
-}
-
-TEST(LinuxHeadersUtils, GetKernelVersionUbuntu) {
-  // Setup: Point to a custom /proc filesystem.
-  PX_SET_FOR_SCOPE(FLAGS_proc_path, GetPathToTestDataFile("sample_host_ubuntu/proc"));
-
-  // Main test.
-  StatusOr<KernelVersion> kernel_version_status = FindKernelVersion(kProcFSKernelVersionSources);
-  ASSERT_OK(kernel_version_status);
-  KernelVersion kernel_version = kernel_version_status.ValueOrDie();
-  EXPECT_EQ(kernel_version.code(), 0x050441);
-}
-
-TEST(LinuxHeadersUtils, GetKernelVersionDebian) {
-  // Setup: Point to a custom /proc filesystem.
-  PX_SET_FOR_SCOPE(FLAGS_proc_path, GetPathToTestDataFile("sample_host_debian/proc"));
-
-  // Main test.
-  StatusOr<KernelVersion> kernel_version_status = FindKernelVersion(kProcFSKernelVersionSources);
-  ASSERT_OK(kernel_version_status);
-  KernelVersion kernel_version = kernel_version_status.ValueOrDie();
-  EXPECT_EQ(kernel_version.code(), 0x041398);
-}
-
-TEST(LinuxHeadersUtils, KernelHeadersDistance) {
-  // Distances should always be positive.
-  EXPECT_GT(KernelHeadersDistance({4, 14, 255}, {4, 15, 10}), 0);
-
-  // Order shouldn't matter.
-  EXPECT_EQ(KernelHeadersDistance({4, 14, 255}, {4, 15, 255}),
-            KernelHeadersDistance({4, 15, 255}, {4, 14, 255}));
-
-  EXPECT_GT(KernelHeadersDistance({4, 14, 1}, {4, 15, 5}),
-            KernelHeadersDistance({4, 14, 1}, {4, 14, 2}));
-
-  // A change in the major version is considered larger than any change in the minor version.
-  EXPECT_GT(KernelHeadersDistance({4, 14, 255}, {4, 15, 0}),
-            KernelHeadersDistance({4, 14, 255}, {4, 14, 0}));
-}
-
-TEST(LinuxHeadersUtils, CompareKernelVersions) {
-  EXPECT_EQ(CompareKernelVersions({1, 2, 3}, {1, 2, 3}), KernelVersionOrder::kSame);
-
-  EXPECT_EQ(CompareKernelVersions({1, 2, 3}, {1, 2, 4}), KernelVersionOrder::kOlder);
-  EXPECT_EQ(CompareKernelVersions({1, 2, 3}, {1, 3, 3}), KernelVersionOrder::kOlder);
-  EXPECT_EQ(CompareKernelVersions({1, 2, 3}, {2, 2, 3}), KernelVersionOrder::kOlder);
-
-  EXPECT_EQ(CompareKernelVersions({1, 2, 4}, {1, 2, 3}), KernelVersionOrder::kNewer);
-  EXPECT_EQ(CompareKernelVersions({1, 3, 3}, {1, 2, 3}), KernelVersionOrder::kNewer);
-  EXPECT_EQ(CompareKernelVersions({2, 2, 3}, {1, 2, 3}), KernelVersionOrder::kNewer);
 }
 
 TEST(LinuxHeadersUtils, ModifyVersion) {
@@ -157,7 +55,7 @@ TEST(LinuxHeadersUtils, ModifyVersion) {
 
   // Functions Under Test
 
-  StatusOr<KernelVersion> host_linux_version = FindKernelVersion();
+  StatusOr<system::KernelVersion> host_linux_version = system::FindKernelVersion();
   ASSERT_OK(host_linux_version);
   uint32_t host_linux_version_code = host_linux_version.ValueOrDie().code();
   EXPECT_GT(host_linux_version_code, 0);
@@ -238,32 +136,37 @@ TEST(LinuxHeadersUtils, FindClosestPackagedLinuxHeaders) {
 #endif
 
   {
-    ASSERT_OK_AND_ASSIGN(PackagedLinuxHeadersSpec match,
-                         FindClosestPackagedLinuxHeaders(kTestSrcDir, KernelVersion{4, 4, 18}));
+    ASSERT_OK_AND_ASSIGN(
+        PackagedLinuxHeadersSpec match,
+        FindClosestPackagedLinuxHeaders(kTestSrcDir, system::KernelVersion{4, 4, 18}));
     EXPECT_THAT(match.path.string(), EndsWith(prefix + "4.14.176.tar.gz"));
   }
 
   {
-    ASSERT_OK_AND_ASSIGN(PackagedLinuxHeadersSpec match,
-                         FindClosestPackagedLinuxHeaders(kTestSrcDir, KernelVersion{4, 15, 10}));
+    ASSERT_OK_AND_ASSIGN(
+        PackagedLinuxHeadersSpec match,
+        FindClosestPackagedLinuxHeaders(kTestSrcDir, system::KernelVersion{4, 15, 10}));
     EXPECT_THAT(match.path.string(), EndsWith(prefix + "4.14.176.tar.gz"));
   }
 
   {
-    ASSERT_OK_AND_ASSIGN(PackagedLinuxHeadersSpec match,
-                         FindClosestPackagedLinuxHeaders(kTestSrcDir, KernelVersion{4, 18, 1}));
+    ASSERT_OK_AND_ASSIGN(
+        PackagedLinuxHeadersSpec match,
+        FindClosestPackagedLinuxHeaders(kTestSrcDir, system::KernelVersion{4, 18, 1}));
     EXPECT_THAT(match.path.string(), EndsWith(prefix + "4.18.20.tar.gz"));
   }
 
   {
-    ASSERT_OK_AND_ASSIGN(PackagedLinuxHeadersSpec match,
-                         FindClosestPackagedLinuxHeaders(kTestSrcDir, KernelVersion{5, 0, 0}));
+    ASSERT_OK_AND_ASSIGN(
+        PackagedLinuxHeadersSpec match,
+        FindClosestPackagedLinuxHeaders(kTestSrcDir, system::KernelVersion{5, 0, 0}));
     EXPECT_THAT(match.path.string(), EndsWith(prefix + "5.3.18.tar.gz"));
   }
 
   {
-    ASSERT_OK_AND_ASSIGN(PackagedLinuxHeadersSpec match,
-                         FindClosestPackagedLinuxHeaders(kTestSrcDir, KernelVersion{5, 7, 20}));
+    ASSERT_OK_AND_ASSIGN(
+        PackagedLinuxHeadersSpec match,
+        FindClosestPackagedLinuxHeaders(kTestSrcDir, system::KernelVersion{5, 7, 20}));
     EXPECT_THAT(match.path.string(), EndsWith(prefix + "5.3.18.tar.gz"));
   }
 }
